==> 4. Project - Python/03 Scripts/2. Starting with Jupyter and Python.ipynb <==
# 01. Importing Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os

## 02. Working with data types

# Working with data types

print("\nDirection 7:")

for x in range(1,4):
    a = random.randint(1, 99)
    b = 100-a
    c = a + b
    print("%d) %d + %d = %d" % (x, a, b, c))



print("\nDirection 8:")
    
for x in range(1,2):
    d = random.uniform(1.0, 10.0)
    e = random.uniform(10.0, 100.0)
    f = e / d
    print("%d) %s / %s = %s" % (x, e, d, f))



print("\nDirection 9:")

g = "Hel"
h = "lo"
i = "Wo"
j = "rld"

print(g + h + ", " + i + j + "!")



print("\nDirection 10:")

k = g + h
l = i + j
m = ", "

n = "What"
o = "a"
p = "great"
q = "day"
r = "to"
s = "learn"
t = "Python"

print(k + m + l + "!")
print(n + " " + o + " " + p + " " + q + " " + r + " " + s + " " + t +"!")



print("\nBonus Task:")

u = "This string will appear 100 times! "
print(100*u)
