==> 4. Project - Python/03 Scripts/3. Starting with Pandas.ipynb <==
# 01. Importing libraries

# Importing libraries

import pandas as pd
import numpy as np
import os


# 02. Main Tasks

# Direction 4

path = r'X:\Data_Analytic\JUPYTER\Instacart Basket Analysis'


# Direction 5

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df_orders = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)


# Direction 6

df_orders.head()


# Direction 7

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)


# Direction 8.1

df_prods.head(20)


# Direction 8.2

df_prods.tail(30)


#Direction 9

df_prods.columns


#Direction 10

df_prods.shape


df_orders.shape

#Direction 11.1

df_prods.describe()


#Direction 11.2

df_prods.dtypes

#Direction 11.3

The maximum value of the "aisle_id" column is 134.

The data type of the "department_id" column is "int64" (integer).
