==> 4. Project - Python/03 Scripts/4. Starting with Data Wrangling & Subsetting.ipynb <==
# 01 Importing libraries & Data Sets

# Import Libraries

import pandas as pd
import numpy as np
import os


# Import Data Sets

path = r'C:\JUPYTER\Instacart Basket Analysis'

#vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df_test = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df_orders = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)


# 02 Working on the data sets

# Dropping Columns

df_orders.drop(columns = ['eval_set'])


df_test['days_since_prior_order'].value_counts(dropna = False)

df_test.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_test.head()

df_test.describe()

df_test['order_id'] = df_test['order_id'].astype('str')

df_test['order_id'].dtype

df_dep.head()

df_dep.T

df_dep_t = df_dep.T

df_dep_t

df_dep_t.reset_index()

new_header = df_dep_t.iloc[0]

new_header

df_dep_t_new = df_dep_t[1:]

df_dep_t_new

df_dep_t_new.columns = new_header

df_dep_t_new

data_dict = df_dep_t_new.to_dict('index')

data_dict

df_snacks = df_prods[df_prods['department_id']==19]

df_snacks.head()
