==> 4. Project - Python/03 Scripts/4. Working with Data Wrangling & Subsetting.ipynb <==
# 01 Importing libraries

# Importing libraries

import pandas as pd
import numpy as np
import os


## 01.1 Importing Data Sets

# Importing Data Sets

path = r'X:\Data_Analytic\JUPYTER\Instacart Basket Analysis'

df_orders = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)




# 02 Direction 1 - Working on the tasks from the material

# Dropping eval_set

df_orders_dropped_eval = df_orders.drop(columns = ['eval_set'])

df_orders_dropped_eval

#Count values with NA

df_orders['days_since_prior_order'].value_counts(dropna = False)

# Changing Column Name

df_orders_dropped_eval.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_orders_dropped_eval.head()

# Changing Order_id Into String

df_orders_dropped_eval['order_id'] = df_orders_dropped_eval['order_id'].astype('str')

df_orders_dropped_eval['order_id'].dtype


df_orders_dropped_eval.describe()

# Testing df_dep

df_dep.head()

# Switching Rows and Columns of df_dep

df_dep.T

# Change Header df_dep

df_dep_t = df_dep.T # saves the switch df version on a new variable

df_dep_t.reset_index() # resets index on that new df

new_header = df_dep_t.iloc[0] # saves the new headers under a new variable

df_dep_t_new = df_dep_t[1:] # saves the new df starting at row 1 under new variable

df_dep_t_new.columns = new_header # renames the columns from the new df

df_dep_t_new

# Creating the Data Dictionary

data_dict = df_dep_t_new.to_dict('index')

data_dict

# Looking at the Snacks

df_snacks = df_prods.loc[df_prods['department_id'].isin([19])]

df_snacks.head()

# 03 - Working on the Tasks from the Directions

## Direction 2

df_orders_dropped_eval['user_id'] = df_orders_dropped_eval['user_id'].astype('str')

df_orders_dropped_eval['user_id'].dtype

## Direction 3

# Note, that I already saved the original data frame under a new variable - and chose this specific column, because
# I think that all columns already had descriptive names in the first place. However, now it's clear that the time
# describes the 24 hour format

df_orders_dropped_eval.rename(columns = {'order_hour_of_day' : 'order_hour_of_day_24hrs'}, inplace = False)

## Direction 4

df_orders_dropped_eval['order_hour_of_day'].value_counts(dropna = False)

# The busiest hour for placing orders is the 10th hour of the day

## Direction 5

print(data_dict.get('4'))

# The meaning behind the value of 4 is 'produce'

## Direction 6

df_breakfast = df_prods.loc[df_prods['department_id'].isin([14])]

df_breakfast

## Direction 7

df_dinnerparties = df_prods.loc[df_prods['department_id'].isin([5,20,7,12])]

df_dinnerparties

## Direction 8

df_dinnerparties.info()

# The Dataframe 'df_dinnerparties' contains 7650 rows (as can be seen below and in Direction 7)

## Direction 9

df_user_one = df_orders_dropped_eval.loc[df_orders_dropped_eval['user_id']=='1']

df_user_one

## Direction 10

df_user_one.describe()

df_user_one.info()

# He/She placed 11 Orders, most of them on a thursday, he/she never ordered something on the 
# weekend (including friday). He/She mostly orders in the morning and has rather large days since the last order as 
# well, therefore he/she doesn't order often. He/She has placed 11 orders so far. The data set containing his behavior
# is around 616 bytes big.

# 03 Exporting the dataframes

# Exporting the dataframes

df_orders_dropped_eval.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'))

df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))

df_user_one.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'user_1.csv'))

df_dinnerparties.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'dinnerparties.csv'))

df_breakfast.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'breakfast.csv'))
