==> 4. Project - Python/03 Scripts/5. Data Consistency Checks.ipynb <==
# 01 Importing Libraries

# Importing Libraries

import pandas as pd
import numpy as np
import os


## 01.1 Importing Data Frames

# Importing Data Frames

path = r'C:\JUPYTER\Instacart Basket Analysis'

#df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)
#df_ords2 = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)
df_ords2 = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)


df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

#df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'product'))

# 02 Working on the tasks from the material

# Create a dataframe

df_test = pd.DataFrame()

# Create a mixed type column

df_test['mix'] = ['a', 'b', 1, True]

df_test

# Check columns for mixed types

for col in df_test.columns.tolist():
    weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_test[weird]) > 0:
        print (col)

df_test['mix'] = df_test['mix'].astype('str')

# Finding Missing Values

df_prods.isnull().sum()


df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.shape

# Filtering the Missing values and saving all under new variable

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

# Looking for duplicates

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

## Exporting the Dataframe from within the materials task

# Exporting the Dataframe from within the materials task

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

# 03 Working on the Directions from the task

## Direction 2

#df_prods.describe()

df_prods_clean_no_dups.describe()

I think the prices need to be investigated, the max value of '99999' seems odd.

## Direction 3

#Checking for mixed data types

for colu in df_ords.columns.tolist():
    odd = (df_ords[[colu]].applymap(type) != df_ords[[colu]].iloc[0].apply(type)).any(axis = 1)
    if len (df_ords[odd]) > 0:
        print (colu)
        
# I checked 3 times, even with different variable names (as seen above), but I don't get any mixed data type columns.
# I even closed all jupyter notebooks (with close and halt), but I still don't get any odd results.



## Direction 5

# Checking for missing values

df_ords.isnull().sum()

I already noted this on the behavior of user number 1 from last excercise. The first order was 'NaN', so I think that the first order is always 'NaN', because how can there be any day prior the last order, if it's the first?

## Direction 6

# Addressing the missing values

zero = -1

df_ords_nan = df_ords

df_ords_nan = df_ords_nan['days_since_prior_order'].fillna(zero, inplace = True)



#Testing what the output for days_since_prio_order = 0 is

df_ords_sub = df_ords.loc[df_ords['days_since_prior_order'].isin([-1])]

df_ords_sub

#Testing what the normal output for days_since_prio_order = 0 is before change, using a second imported data frame

df_ords_not_sub = df_ords2.loc[df_ords2['days_since_prior_order'].isin([0])]

df_ords_not_sub

I introduced a new variable, zero with the value 0 and then filled the missing values within the dataset with those zeros.
Afterwards I checked if the count of zeros is corrected and panicked for a second, because it's not. But then I remembered
that user number 1 also had a day since prior order with the value of 0 meaning he/she ordered two days in a row.

273964 - 67755 = 206209

#### EDIT:

Now I, instead of replacing the missing values with 0 i replaced them with '-1', as a sign, that the user is ordering for the first time. I could also add another column, etc. But replacing the missing ones with 0 might be misleading. Although the replacement with '-1' might be misleading as well, but I still think it's better than 0.

## Direction 7

# Checking df_ords for duplicates

df_ords_dups = df_ords[df_ords.duplicated()]

df_ords_dups




df_ords_summing = df_ords.duplicated().sum()

df_ords_summing


df_ords_summing_test = df_ords['days_since_prior_order'].duplicated().sum()

df_ords_summing_test

df_ords.shape

df_ords.tail()

For direction 7 I used 2 things once the method we learned, then a summing method and a test, if my summing method actually worked an a single column alone. So far, I can't say that I've found duplicates. Maybe I made something wrong.

Anyway, I exported the csv, loaded it into excel to at least have a quick check for the firs 'few' rows. I didn't find any duplicates in excel as well.

## 03.1 Exporting the dataframes

# Exporting the data frames

df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'))
