==> 4. Project - Python/03 Scripts/6. Combining and Exporting Data.ipynb <==
# Working on the tasks of 4.6

#### 1. Import Libraries
#### 2. Import DFs
#### 3. Workin on the Directions
#### 4. Exporting the Dataframe

## 1. Import Libraries

# Import Libraries

import pandas as pd
import numpy as np
import os


## 2. Import DFs

# Import Dataframes

path = r'X:\Data_Analytic\JUPYTER\Instacart Basket Analysis'

df_ords_prods_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))


df_prods_checked = pd.read_csv(os.path.join( path, '02 Data', 'Prepared Data', 'products_checked.csv'))

## 3. Working on the Directions

### Direction 4

df_ords_prods_combined.shape

# The shape is equal.

### Direction 5

df_prods_checked.head()

df_ords_prods_combined.head()

##### Explanation, the '-1' in the 'days_since_prior_order' stem from a correction I did, before receiving my results on the last task. In the meantime, I realised, that giving the NaN a 0 could indicate the wrong information. I want to make clear, that this is the customers very first order, by changing the NaN to a '-1'. So I corrected it, but thought that maybe it isn't that big of a problem and did not resubmit my correction.

df_ords_prods_combined.tail()

df_ords_prods_combined_dropped = df_ords_prods_combined.drop(['Unnamed: 0', 'Unnamed: 0.1'], axis=1)

##### I believe that due to the merging of the previously dataframes, the 'Unnamed: 0' and 'Unnamed: 0.1' dataframes might be the indexes. Anyway, I dropped them under a new variable.

df_ords_prods_combined_dropped.head()

##### Furthermore I'm going to rename the '_merge' column, so that it's understandable that this column stems from the first merge.

df_ords_prods_combined_dropped.rename(columns={'_merge' : 'first_m'})

df_ords_prods_combined_dropped.head()

##### I also dropped the 'Unnamed: 0' column from the df_prods_checked dataframe

df_prods_checked_dropped = df_prods_checked.drop(['Unnamed: 0'], axis=1)

df_prods_checked_dropped.head()

### Merging the two dataframes

df_merged_inner = df_ords_prods_combined_dropped.merge(df_prods_checked_dropped, on = 'product_id', indicator = 'Second_Merge')

#df_merged_full = df_ords_prods_combined_dropped.merge(df_prods_checked_dropped, on = 'product_id', how = 'outer')

df_merged_full = df_ords_prods_combined_dropped.merge(df_prods_checked_dropped, on = 'product_id', how = 'outer', indicator = 'Second_Merge')

df_merged_inner.head()

df_merged_full.head()

### Direction 6

df_merged_full['Second_Merge'].value_counts()

df_merged_inner['Second_Merge'].value_counts()

##### As I checked the full_merge data frame I wanted to know what I'm dealing with exactly, so I created 2 subsets and looked at them

df_merged_full_left = df_merged_full[df_merged_full['Second_Merge'] == 'left_only']

df_merged_full_right = df_merged_full[df_merged_full['Second_Merge'] == 'right_only']

df_merged_full_left

df_merged_full_right

##### As we can see, there probably are 11 products which haven't been ordered at all or are mistakes and there are probably 30200 customers who ordered something that doesn't exist in the database or are mistakes as well. But since these mistakes are so little, I will ignore them and use the inner joined dataframe.

## 4. Exporting the Dataframe / Direction 7

df_merged_inner.shape

df_merged_inner.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
