==> 4. Project - Python/03 Scripts/7. Deriving New Variables.ipynb <==
# Structure

### 1. Importing Libraries
### 2. Importing Dataframes
### 3. Defining Functions
### 4. If-Statements with For-Loops
### 5. Workin through the Directions of the task

## 1. Importing Libraries

# Importing Libraries

import pandas as pd
import numpy as np
import os

## 2. Importing Dataframes

# Importing Dataframes

path = r'X:\Data_Analytic\JUPYTER\Instacart Basket Analysis'

ords_prods_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))


# Selecting the first Million rows of the DF

df = ords_prods_merged[:1000000]

## 3. Defining Functions

#### If-Statements with user-defined function

# Defining the function price label

def price_label(row):

    if row['prices'] <= 5:
        return 'Low-range product'

    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'

    elif row['prices'] > 15:
        return 'High-range product'

    else: return 'Not enough data'

df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

#### If-Statements with the loc() function

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

#### If-Statements with the loc() function on whole dataframe

ords_prods_merged.loc[ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

ords_prods_merged.loc[(ords_prods_merged['prices'] <= 15) & (ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

ords_prods_merged.loc[ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

ords_prods_merged['price_range_loc'].value_counts(dropna = False)

## 4. If-Statements with For-Loops

# First Test of for-loop

for x in range(30,45):
    print("My age is %d" % (x))

# Frequency of 'orders_day_of_week'

ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)


# For Loop Column rating busiest day of week

result = []

for value in ords_prods_merged['orders_day_of_week']:
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy')
    else:
        result.append('Regularly busy')

# Print List

result

# Add result into df

ords_prods_merged['busiest_day'] = result

# Count added result list of df

ords_prods_merged['busiest_day'].value_counts(dropna = False)

## 5. Working through the directions of the task

### Direction 2

##### In order to change the busiest day column to busiest days, i will drop the column, then work again with a list and a for-loop

ords_prods_merged.head()

ords_prods_merged_new = ords_prods_merged.drop('busiest_day', axis=1)

ords_prods_merged_new.head()

ords_prods_merged_new['orders_day_of_week'].value_counts(dropna = False)

# Creating the For-Loop

result = []

for x in ords_prods_merged_new['orders_day_of_week']:
    if x == 0:
        result.append('Busiest day')
    elif x == 1:
        result.append('Busy day')
    elif x == 3:
        result.append('Less busy day')
    elif x == 4:
        result.append('Least busy day')
    else:
        result.append('Regularly busy day')

ords_prods_merged_new['busiest_day'] = result

ords_prods_merged_new['busiest_day'].value_counts(dropna = False)

### Direction 3

##### I checked it and the results are correct.

### Direction 4

This one is a bit longer. All due to the fact, that I thought it would be cool, to have this process kinda automated, meaning that I wrote a script that can be run everytime changes occur and the time_periods and labels automatically work.

So the idea is, to first establish a way of automating the chosing of what are busy hours and what are less busy hours. I decided to use the IQRs which i get from the describe() function. Then I labelled the hours using the IQRs. And at the end merge this back into the main df.

ords_prods_merged_new['order_hour_of_day'].value_counts(dropna = False)

# This code takes the count of the hours and transforms it into a new df. Then it saves the describe function of this df
# and saves it under a nwe dataframe. Then It takes the values for the 25% and 75% quartile and saves those under 
# two variables. And finally prints those to check if it worked.

df_hour_count = ords_prods_merged_new['order_hour_of_day'].value_counts(dropna = False)

df_hour_described = df_hour_count.describe()

first_q = df_hour_described.values[4]

fourth_q = df_hour_described.values[6]

print(first_q)
print(fourth_q)

# Here I take the df with the count of the hours and reset it's index under a new df

df_hour_count_r = df_hour_count.reset_index()

# Here I rename the columns of the df with the resetted index

df_hour_count_r = df_hour_count_r.rename(columns={"index" : "hour_of_day", "order_hour_of_day" : "count"})

# Here I used a function, since the df is so small
#
# This function adds a label if the count is either between, higher or lower than the 25% / 75% quartile.

def label_hours(row):
    if row['count'] <= first_q:
        return 'fewest orders'
    elif (row['count'] > first_q) and (row['count'] <= fourth_q):
        return 'average orders'
    elif row['count'] > fourth_q:
        return 'most orders'
    else: return 'not enough data'


# The function is applied to the df

df_hour_count_r['labels'] = df_hour_count_r.apply(label_hours, axis=1)

# Here I check the DF if it worked properly

df_hour_count_r

amount_most_orders = (df_hour_count_r['labels'] == 'most orders').sum()

amount_average_orders = (df_hour_count_r['labels'] == 'average orders').sum()

amount_fewest_orders = (df_hour_count_r['labels'] == 'fewest orders').sum()

print(amount_most_orders)
print(amount_average_orders)
print(amount_fewest_orders)

list_most_orders = []
list_average_orders = []
list_fewest_orders = []


for i in range(0, amount_most_orders):
    a = df_hour_count_r.loc[df_hour_count_r['labels'] == 'most orders']['hour_of_day'].values[i]
    list_most_orders.append(a)


for ii in range(0, amount_average_orders):
    aa = df_hour_count_r.loc[df_hour_count_r['labels'] == 'average orders']['hour_of_day'].values[ii]
    list_average_orders.append(aa)
    
for iii in range(0, amount_fewest_orders):
    aaa = df_hour_count_r.loc[df_hour_count_r['labels'] == 'fewest orders']['hour_of_day'].values[iii]
    list_fewest_orders.append(aaa)



list_most_orders.sort()
list_average_orders.sort()
list_fewest_orders.sort()


#print(list_most_orders)
#print(list_average_orders)
#print(list_fewest_orders)



    

### Sorting the lists into time-periods

The following 4 codefragments have the simple function of taking the lists and transforming them into time periods, so out of [0, 1, 2, 3, 4, 5] will become ['0-5']

The idea is the following, i go through the list and test if the second result minus the first result is '-1' or to be clear is NOT '-1'. If that's the case I will save the distance between the first and the last number that are on this list so far. And save this as a string into a new list. If I ran through the list from the for-loop I finish the code and save the last distance as well as a string.

The following is the test version I created to check if my work would withstand change in the dataframe

# THE TEST VERSION

test_list = [1, 2, 3, 4, 5, 9, 10, 11, 14, 15, 16, 18, 19, 20]

#print(test_list)

ende = 13

x = 0
y = 1

list_start = 0
list_ende = 1

list_interval = []

for z in range(0, ende):
    
    w = test_list[x] - test_list[y]
    
    x = x + 1
    
    y = y + 1
    
    if w != -1:

        int1 = str(test_list[list_start])
        
               
        int2 = str(test_list[list_ende-1])
        
        int3 = int1 + '-' + int2
        
        list_interval.append(int3)            
            
        list_start = list_ende
        
    
    if z == ende-1:
        
        list_start = test_list[list_start]
        
        int1 = str(list_start)
             
        int2 = str(test_list[list_ende])
        
        int3 = int1 + '-' + int2
        
        list_interval.append(int3)        

    list_ende = list_ende + 1
    

print(list_interval)   


# THE VERSION FOR THE MOST ORDERS

mostx = 0
mosty = 1

most_start = 0
most_ende = 1

list_most_interval = []

for mostz in range(1, amount_most_orders):
    
    mostw = list_most_orders[mostx] - list_most_orders[mosty]
    
    mostx = mostx + 1
    
    mosty = mosty + 1
    
    if mostw != -1:
        
        mostint1 = str(list_most_orders[most_start])
        
        mostint2 = str(list_most_orders[most_ende-1])
        
        mostint3 = mostint1 + '-' + mostint2
        
        list_most_interval.append(mostint3)
        
        most_start = most_ende
        
    if mostz == amount_most_orders-1:
        
        most_start = list_most_orders[most_start]
        
        mostint1 = str(most_start)
        
        mostint2 = str(list_most_orders[most_ende])
        
        mostint3 = mostint1 + '-' + mostint2
        
        list_most_interval.append(mostint3)
        
    most_ende = most_ende + 1
    
print(list_most_interval)

# THE VERSION FOR THE AVERAGE ORDERS

averagex = 0
averagey = 1

average_start = 0
average_ende = 1

list_average_interval = []

for averagez in range(1, amount_average_orders):
    
    averagew = list_average_orders[averagex] - list_average_orders[averagey]
    
    averagex = averagex + 1
    
    averagey = averagey + 1
    
    if averagew != -1:
        
        averageint1 = str(list_average_orders[average_start])
        
        averageint2 = str(list_average_orders[average_ende-1])
        
        averageint3 = averageint1 + '-' + averageint2
        
        list_average_interval.append(averageint3)
        
        average_start = average_ende
        
    if averagez == amount_average_orders-1:
        
        average_start = list_average_orders[average_start]
        
        averageint1 = str(average_start)
        
        averageint2 = str(list_average_orders[average_ende])
        
        averageint3 = averageint1 + '-' + averageint2
        
        list_average_interval.append(averageint3)
        
    average_ende = average_ende + 1
    
print(list_average_interval)

# THE VERSION FOR THE FEWEST ORDERS

fewestx = 0
fewesty = 1

fewest_start = 0
fewest_ende = 1

list_fewest_interval = []

for fewestz in range(1, amount_fewest_orders):
    
    fewestw = list_fewest_orders[fewestx] - list_fewest_orders[fewesty]
    
    fewestx = fewestx + 1
    
    fewesty = fewesty + 1
    
    if fewestw != -1:
        
        fewestint1 = str(list_fewest_orders[fewest_start])
        
        fewestint2 = str(list_fewest_orders[fewest_ende-1])
        
        fewestint3 = fewestint1 + '-' + fewestint2
        
        list_fewest_interval.append(fewestint3)
        
        fewest_start = fewest_ende
        
    if fewestz == amount_fewest_orders-1:
        
        fewest_start = list_fewest_orders[fewest_start]
        
        fewestint1 = str(fewest_start)
        
        fewestint2 = str(list_fewest_orders[fewest_ende])
        
        fewestint3 = fewestint1 + '-' + fewestint2
        
        list_fewest_interval.append(fewestint3)
        
    fewest_ende = fewest_ende + 1
    
print(list_fewest_interval)

# JOINING THE LISTS AS A STRING

fewest_time_period = ', '.join(list_fewest_interval)

average_time_period = ', '.join(list_average_interval)

most_time_period = ', '.join(list_most_interval)


print(fewest_time_period)
print(average_time_period)
print(most_time_period)



### Now I will add everything to the main dataframe

for i in range(0, amount_most_orders):
    ords_prods_merged_new.loc[ords_prods_merged_new['order_hour_of_day'] == list_most_orders[i], 'busiest_period_of_day'] = 'Most orders'
    

for i in range(0, amount_average_orders):
    ords_prods_merged_new.loc[ords_prods_merged_new['order_hour_of_day'] == list_average_orders[i], 'busiest_period_of_day'] = 'Average orders'

for i in range(0, amount_fewest_orders):
    ords_prods_merged_new.loc[ords_prods_merged_new['order_hour_of_day'] == list_fewest_orders[i], 'busiest_period_of_day'] = 'Fewest orders'

### Now I test, if the values are correct. / Direction 5

ords_prods_merged_new['busiest_period_of_day'].value_counts(dropna = False)

df_hour_count_r.loc[df_hour_count_r['labels'] == 'most orders', 'count'].sum()

df_hour_count_r.loc[df_hour_count_r['labels'] == 'average orders', 'count'].sum()

df_hour_count_r.loc[df_hour_count_r['labels'] == 'fewest orders', 'count'].sum()

### Now I add the time periods.

ords_prods_merged_new.loc[ords_prods_merged_new['busiest_period_of_day'] == 'Most orders', 'time_period'] = most_time_period

ords_prods_merged_new.loc[ords_prods_merged_new['busiest_period_of_day'] == 'Average orders', 'time_period'] = average_time_period

ords_prods_merged_new.loc[ords_prods_merged_new['busiest_period_of_day'] == 'Fewest orders', 'time_period'] = fewest_time_period

## Exporting the main dataframe

ords_prods_merged_new.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_time_period_combined.pkl'))
