==> 4. Project - Python/03 Scripts/9. Starting with Matplotlib and Seaborn.ipynb <==
# Stucture

### 1. Import libraries and setting up library options
### 2. Import dataframes
### 3. Checking the dataframe
### 4. Merging the dataframes
### 5. Exporting the dataframe

## 1. Import libraries and setting up library options

# Importing libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

plt.rcParams['figure.figsize']=(12.8, 7.2) # Increases the chart size to 1280x720 px
# I just finished part 1 and realized, I don't need to change the size.

## 2. Import dataframes

path = r'X:\Data_Analytic\JUPYTER\Instacart Basket Analysis'

cust_df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'), index_col = False)

ords_prods_df = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_flagged.pkl'))

## 3. Checking the dataframe

ords_prods_df.shape

cust_df.head()

cust_df.tail()

cust_df.shape

cust_df.describe()

cust_df.info()

After seeing these informations, I assume, that there is missing data in the 'First Name' column. Furthermore, the quantifiable columns seems to be alright. And the data types are correct as well.

I'm thinking about changing the Null values to something like 'Name not given', assuming that these users might not trust the instacart shop not enough to give it their name. But then again, if I need this info I could work on it in the future as well without any real timeloss as it doesn't seem that important for the further tasks. So I will just drop the column.

cust_df_dropped = cust_df.drop(['First Name'], axis=1)

cust_df_dropped.head()

cust_df_dropped.info()

# Now I will change the columns as well to fit to the other df's

cust_df_dropped = cust_df_dropped.rename(columns = {'Surnam': 'surname', 'Gender': 'gender', 'STATE': 'state', 'Age': 'age'})

cust_df_dropped.head()

##### Checking consistency

# Data Type issues

for col in cust_df_dropped.columns.tolist():
    weird = (cust_df_dropped[[col]].applymap(type) != cust_df_dropped[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (cust_df_dropped[weird]) > 0:
        print (col)

# Missing Values

cust_df_dropped.isnull().sum()

# Checking duplicates

df_dups = cust_df_dropped[cust_df_dropped.duplicated()]

df_dups

cust_df_dropped.shape

## 4. Merging the dataframes

cust_df_dropped.head()

pd.set_option('display.max_columns', None)

ords_prods_df.head()

ords_prods_df.describe()

I just checked to make sure if merging on user_id would involve difficulties, but everything looks alright to merge.

ords_prods_df.info()

cust_df_dropped.info()

##### The User_id has the same data type

# Merging the dataframes

df_merged = ords_prods_df.merge(cust_df_dropped, on = 'user_id', indicator = 'third_merge')

df_merged.head()

# Dropping the merge indicators

df_merged_dropped = df_merged.drop(['_merge', 'Second_Merge', 'third_merge'], axis=1)

df_merged_dropped.head()

df_merged_dropped.shape

df_merged.shape

## 5. Exporting the dataframe

df_merged_dropped.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'full_merged_4-9-p1.pkl'))
